some refactoring + comments
