+++ pages/api/resources/[resourceId].ts
@@ -2,15 +2,7 @@ import type { NextApiRequest, NextApiResponse } from 'next'
 import type {Resource} from '../../../models/types'
 import FirebaseInteractor from '../../../firebase/firebaseInteractor'
 import { DocumentData, FirestoreDataConverter, QueryDocumentSnapshot } from 'firebase/firestore'
- 
-const resourceConverter : FirestoreDataConverter<Resource> = {
-    toFirestore: (data: Resource) : DocumentData => data,
-    fromFirestore: (snap: QueryDocumentSnapshot) : Resource => {
-      let data = snap.data() as Resource
-      data.id = snap.id
-      return data
-    }
-}
+import { resourceConverter } from '../../../firebase/converters'
 
 export default async function handler(
   req: NextApiRequest,
